+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - A bunch of icon corrections for effects in v12. @callbritton
 - Add character class to spells when importing from a character from DDB.
 - Conjure Spells will now use the CR conjuring system.
+- Find Familiar uses summoning system.
 
 # 5.2.25